# backend/openapi.yaml
openapi: 3.1.0
info:
  title: 補助金申請書自動生成API
  version: 1.0.0
  description: 補助金申請書類の自動生成、検証、エクスポートを行うAPI
  contact:
    name: API Support
    email: support@subsidy-app.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-api-version-info:
    deprecated: false
    sunset: null
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.subsidy-app.example.com
    description: Production server
...
paths:
...
  /draft:
    post:
      tags:
        - draft
      summary: 事業計画の草案を一括生成
      operationId: createDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftRequest'
      responses:
        '200':
          description: 草案生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /validate/text:
    post:
      tags:
        - validate
      summary: テキスト検証
      operationId: validateText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
                - rules
              properties:
                content:
                  type: string
                rules:
                  type: array
                  items:
                    type: string
                    enum: [char_count, forbidden_words, redundancy]
      responses:
        '200':
          description: 検証完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /validate/plan:
    post:
      tags:
        - validate
      summary: 事業計画の整合性・様式チェック
      operationId: validatePlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanValidateRequest'
      responses:
        '200':
          description: 検証完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /validate/kpis:
    post:
      tags:
        - validate
      summary: KPI検証
      operationId: validateKpis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kpis
              properties:
                kpis:
                  type: array
                  items:
                    $ref: '#/components/schemas/Kpi'
      responses:
        '200':
          description: 検証完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
...
components:
  schemas:
    DraftRequest:
      type: object
      required: [scheme_id, goal]
      properties:
        scheme_id:
          type: string
          description: 対象制度ID（例 jizokuka-2025-v1）
        goal:
          type: string
          description: 何をしたいか（要約）
        constraints:
          type: object
          properties:
            budget_max:
              type: number
            deadline:
              type: string
              format: date
            staff:
              type: integer
        market:
          type: object
          properties:
            region:
              type: string
            industry:
              type: string
        assets:
          type: array
          items:
            type: string
    DraftResponse:
      type: object
      properties:
        draft_id:
          type: string
        sections:
          type: object
          properties:
            background:
              type: string
            problem_solution:
              type: string
            plan_5w1h:
              type: object
              properties:
                what: { type: string }
                why: { type: string }
                how: { type: string }
                who: { type: string }
                when: { type: string }
                where: { type: string }
            kpi:
              type: array
              items:
                $ref: '#/components/schemas/Kpi'
            roadmap:
              type: array
              items:
                type: object
                properties:
                  milestone: { type: string }
                  start: { type: string, format: date }
                  end: { type: string, format: date }
            budget:
              type: array
              items:
                type: object
                properties:
                  category: { type: string }
                  amount: { type: number }
            team:
              type: array
              items:
                type: object
                properties:
                  role: { type: string }
                  member: { type: string }
            risks:
              type: array
              items:
                type: object
                properties:
                  risk: { type: string }
                  mitigation: { type: string }
            differentiation:
              type: string
        references:
          type: array
          items:
            type: object
            properties:
              title: { type: string }
              source: { type: string }
              url: { type: string, format: uri }
...
    PlanValidateRequest:
      type: object
      required: [scheme_id, text]
      properties:
        scheme_id:
          type: string
        text:
          type: string
        numbers:
          type: object
          properties:
            budget_total: { type: number }
            subsidy_rate: { type: number }
        structure:
          type: object
          properties:
            sections:
              type: array
              items: { type: string }
...
    Kpi:
      type: object
      required:
        - name
        - baseline
        - target
        - unit
      properties:
        name:
          type: string
        baseline:
          type: number
        target:
          type: number
        unit:
          type: string
        measurement_method:
          type: string
        frequency:
          type: string
...
    ValidationResponse:
      type: object
      properties:
        is_valid:
          type: boolean
        errors:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              message:
                type: string
              field:
                type: string
              severity:
                type: string
                enum: [error, warning, info]
        suggestions:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              message:
                type: string
              proposed_fix:
                type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: リクエスト不正
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []